sequence: Reverted the SeqString comparison to what it was a while back.
Comparison first checks the structure of the SeqString again before comparing numbers.
Behavior is now identical to what it was a while back (before May 2010).
Before this patch, the string "P1z" was smaller than "P2a" which seems correct at first
glance but this may mess with sequences. If there is a "P2z" it will come after "P2a"
which is not desirable anymore. Another example is a folder of image files *.JPG and *.CR2
that have the same names but only differ in extension. Before this patch, seqls wouldn't
display two sequences because the files get sorted interleaved.
With this patch, the string parts have precedence and the numbers are only compared if the
string structure is identical.