sequence: Added the signedNums argument to SeqString constructor, buildSequences() and glob().
That argument can be used to enable support for signed numbers inside strings. If a number
is treated as a signed number, any preceding minus sign is assumed to be part of the number
instead of the preceding string.
(the command line utilities haven't been updated yet, they still assume unsigned numbers)